Working on the player UI - adding support of a seeking position of a track: track bar click, accumlating arrowns keys - similar support of a volume change - fiound a bug, that frequent chnage of an UI asserts in FAR; added mutex to the semd_dlg_msg function  - since the caches control now their own resync timings, there is no need to have sync thread frequence as 1Hz, changed to ~20Hz and added specific events, so other classes can execute their code in the aux thread

+++ src/ui/views.cpp
@@ -160,7 +160,7 @@ namespace spotifar
             for (auto& [id, track]: api.get_tracks(album_id))
             {
                 std::wstring track_name = std::format(L"{:02}. {}", track.track_number, track.name);
-                result.push_back({id, track_name, L"", TRACK_ITEM_ATTRS, track.duration});
+                result.push_back({id, track_name, L"", TRACK_ITEM_ATTRS, (size_t)track.duration});
             }
             return result;
         }
